refactor(anthropic-direct): extract pure param resolvers into resolve-params.ts
Move the cluster of stateless, instance-independent parameter resolvers out
of the 1,135-LOC provider file into a focused module, shrinking index.ts by
~195 lines and isolating independently testable pure logic.

Moved to src/agent/providers/anthropic-direct/resolve-params.ts:
  - resolveAutoCompactThreshold
  - resolveMaxTokens
  - resumeHistoryToMessages
  - resolveThinkingParam
  - resolveEffort
  - warnedTokenClamps (shared per-process dedupe Set used by resolveMaxTokens
    and resolveThinkingParam — moved to preserve warn-once behavior)
  - supporting privates the above depend on: isOpus47Plus,
    DEFAULT_AUTO_COMPACT_THRESHOLD, THINKING_OUTPUT_RESERVE_FRACTION

isOpus47Plus is used only by resolveThinkingParam, so it moves too — this
keeps resolve-params.ts free of any back-import from index.ts (no circular
dependency). index.ts re-imports all five resolvers for internal use and
re-exports resolveMaxTokens/resolveThinkingParam/resolveEffort so the
historical `from './index.js'` import path stays valid. The dependent test
(resolve-effort.test.ts) now imports from './resolve-params.js'.

Verification:
  - pnpm lint (tsc --noEmit): pass
  - pnpm audit:sdk:check: pass (no new tracked symbols; the moved
    @anthropic-ai/sdk/resources type imports are not in TRACKED_PACKAGES)
  - resolve-effort.test.ts: 22/22 pass
  - full suite: 2 pre-existing failures only (compact() model-alias and a
    config tier-bindings default), both confirmed present on the base commit
    with these changes stashed — unrelated to this refactor

Closes #103
